Use WakeLocks to keep the CPU running
Recent version of Android are very aggressive about
putting the CPU in a sleep state to optimize battery life,
to the point that locking the screen might cause timers
to finish *minutes* later than expected.

Since a timer that doesn't fire when it's supposed to is
of little use to anyone, acquire a WakeLock whenever a
countdown is running so that the CPU will stay awake, even
if it means burning through the battery faster.